Kernel made compilable for x86_64 target.
git-svn-id: svn://127.0.0.1@180 6511da6e-0655-744e-9e86-4527dd187af8

+++ atomic/kernel/dev/chr/KbdConsole.h
@@ -37,7 +37,7 @@ public:
 	virtual ~KbdConsole();
 
 	virtual int SetInputDevice(KbdDev *dev);
-	virtual u32 GetWaitChannel(Operation op);
+	virtual uintptr_t GetWaitChannel(Operation op);
 	virtual IOStatus Getc(u8 *c);
 	virtual IOStatus OpAvailable(Operation op);
 };
